extract executeTestLater method to ContextStateCollector
